Enhance Solidity contract compilation process and update documentation
- Updated the Makefile to optimize the Solidity contract compilation process, ensuring the use of solcjs and solc with appropriate flags for better performance and compatibility.
- Added detailed echo statements during compilation to inform users about the specific contracts being compiled and any dependencies on pre-compiled artifacts.
- Expanded documentation in ADDING_NEW_MSG_TYPES.md and PROTO_ADDITION_GUIDE.md to include steps for updating precompile support for new message types, ensuring EVM-callable compatibility.
- Introduced a new ERC3643Template contract to standardize security token implementation, mapping ERC-3643 functions to the BitBadges tokenization precompile.
- Updated various contracts and libraries to utilize uint64 for ownership times, aligning with internal BitBadges limits and preventing potential overflow issues.

+++ contracts/ERC3643Tokenization.sol
@@ -19,9 +19,10 @@ contract ERC3643Tokenization is IERC3643 {
     UintRange private constant TOKEN_IDS = UintRange({start: 1, end: 1});
     
     // Ownership times "All" = 1 to MaxUint64
+    // Note: BitBadges uses uint64 internally, so we must use type(uint64).max, not uint256
     UintRange private constant OWNERSHIP_TIMES = UintRange({
         start: 1,
-        end: type(uint256).max
+        end: type(uint64).max
     });
     
     // Reference to the tokenization precompile
